move CorElementType to shared
